+++ lib/node/actions/tool_cache.dart
@@ -8,11 +8,7 @@ import 'dart:js_interop';
 external ToolCache get toolCache;
 
 @JS()
-inline class ToolCache {
-  final JSObject obj;
-
-  ToolCache(this.obj);
-
+extension type ToolCache (JSObject obj) {
   /// Finds the path to a tool version in the local installed tool cache.
   ///
   /// @param toolName      name of the tool